pyproject.toml
[tool.poetry]
name = "adversarial-dbm-gui"
version = "0.1.0"
description = ""
authors = ["Alister Machado dos Reis <a.machadodosreis@uu.nl>"]
readme = "README.md"
packages = [{include = "adversarial_dbm_gui"}]

[tool.poetry.scripts]
main = "adversarial_dbm_gui.main:main"

[tool.poetry.dependencies]
python = "^3.9"
numpy = ">=1.23.5,<1.24.0"
matplotlib = "^3.7.1"
scikit-learn = "^1.2.2"
pandas = "^1.5.3"
multicoretsne = "^0.1"
tqdm = "^4.65.0"
torch = "^2.0.0"
umap-learn = "^0.5.3"
torchvision = "^0.15.1"
python-dotenv = "^1.0.0"
scikit-image = "^0.20.0"
core-adversarial-dbm = {path = "core_adversarial_dbm", develop = true}


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
ipykernel = "^6.22.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
